Device53 VS Удаление предупреждений

+++ sources/S8-53/src/Menu/MenuFunctions.cpp
@@ -56,7 +56,7 @@ void Menu::SetCurrentItem(const void *item, bool active)
         }
         else
         {
-            for(int i = 0; i < NumItemsInPage(page); i++)
+            for(int8 i = 0; i < NumItemsInPage(page); i++)
             {
                 if(Item(page, i) == item)
                 {
